requirements.txt
alembic==1.10.3
click==8.1.3
colorama==0.4.6
decorator==5.1.1
dnspython==2.3.0
email-validator==2.0.0.post2
Flask==2.2.3
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Migrate==4.0.4
flask-paginate==2022.1.8
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
mysqlclient==2.1.1
python-slugify==8.0.1
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.9
SQLAlchemy-Utils==0.41.1
text-unidecode==1.3
typing_extensions==4.5.0
validators==0.20.0
Werkzeug==2.2.3
WTForms==3.0.1
WTForms-Alchemy==0.18.0
WTForms-Components==0.10.5
WTForms-SQLAlchemy==0.3
